fix: rename GameObjectBundle to TiledContentBundle
